Add auditing to Record dictionary
This update adds functionality to the Record object to allow for logging
all changes made to the underlying dictionary. After initialization any
changes can be retrieved as a list of JSON Patch (RFC 6902) operations.
The logging happens when the flush() method is called. This method will
also persist the changes to the Record and any further changes can be
handled the same way.

There are a few caveats. The flush() method has to be manually called.
It should be called when the Record is saved to ArchivesSpace. Also,
because JSON Patch is being used to determine changes to the dictionary,
any additions to a Record must be JSON serializable. Since Records are
populated from JSON and ultimately serialized back to JSON when they are
sent to ArchivesSpace this shouldn't be a problem.